partway in refaqctor. next get itemcontrols out of banneredit
