+++ src/features/GlobalOpts/GlobalOpts.tsx
@@ -6,24 +6,11 @@ export function GlobalOpts() {
 
     return (
         <div className="flex items-center gap-4">
-            master viewport:
 
-            <div className="grid gap-2">
-
-                <input
-                    type="number"
-                    className="border border-gray-400 px-2 w-8"
-                    value={state.masterViewport?.paddingTop ?? ""}
-                    onChange={(e) =>
-                        dispatch({
-                            type: "masterViewport/paddingTop",
-                            payload: { padding: e.target.value },
-                        })
-                    }
-                /> %
+no use for this yet
 
 
             </div>
-        </div>
+      
     );
 };
